requirements.txt
kafka-python
flask
pyspark
pandas
kagglehub